pyproject.toml
[project]
name = "cc-qa-bot-azureai"
version = "0.1.0"
description = "Q&A bot for Confluence Cloud using a GenAI RAG approach with Azure AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atlassian-python-api>=3.41.19",
    "azure-ai-inference>=1.0.0b8",
    "beautifulsoup4>=4.13.3",
    "fastapi[all]>=0.115.8",
    "jinja2>=3.1.5",
    "langchain>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.13.2",
    "requests>=2.32.3",
    "toml>=0.10.2",
    "uuid>=1.30",
    "uvicorn>=0.34.0",
    "wrapt-timeout-decorator>=1.5.1",
]